+++ include/seqan3/utility/views/to.hpp
@@ -12,24 +12,9 @@
 
 #pragma once
 
-#include <seqan3/std/ranges>
+#include <seqan3/utility/range/to.hpp>
 
-#include <seqan3/core/platform.hpp>
-
-#include <range/v3/range/conversion.hpp>
-
-namespace seqan3::views
-{
-
-/*!\brief A to view
- * \ingroup utility_views
- * \details
- * \noapi{This is currently range-v3's to implementation.}
- */
-#if SEQAN3_DOXYGEN_ONLY(1) 0
-inline constexpr auto to;
-#else  // ^^^ doxygen only / real import vvv
-using ::ranges::to;
-#endif // SEQAN3_DOXYGEN_ONLY(1)0
-
-} // namespace seqan3::views
+// clang-format off
+SEQAN3_DEPRECATED_HEADER(
+    "This header is deprecated and will be removed in SeqAn-3.3.0; Please #include <seqan3/utility/range/to.hpp> instead.")
+// clang-format on