+++ test/performance/alphabet/views/view_translate_1D_benchmark.cpp
@@ -14,7 +14,7 @@
 #include <seqan3/alphabet/views/translate.hpp>
 #include <seqan3/test/performance/sequence_generator.hpp>
 #include <seqan3/test/seqan2.hpp>
-#include <seqan3/utility/views/to.hpp>
+#include <seqan3/utility/range/to.hpp>
 
 #ifdef SEQAN3_HAS_SEQAN2
 #    include <seqan/seq_io.h>
@@ -48,7 +48,7 @@ void sequential_read(benchmark::State & state)
     if constexpr (std::is_same_v<tag_t, baseline_tag>)
     {
         seqan3::aa27_vector translated_aa_sequence =
-            dna_sequence | seqan3::views::translate_single | seqan3::views::to<seqan3::aa27_vector>;
+            dna_sequence | seqan3::views::translate_single | seqan3::ranges::to<seqan3::aa27_vector>();
         sequential_read_impl(state, translated_aa_sequence);
     }
     else if constexpr (std::is_same_v<tag_t, translate_tag>)
@@ -89,7 +89,7 @@ void random_access(benchmark::State & state)
     if constexpr (std::is_same_v<tag_t, baseline_tag>)
     {
         seqan3::aa27_vector translated_aa_sequence =
-            dna_sequence | seqan3::views::translate_single | seqan3::views::to<seqan3::aa27_vector>;
+            dna_sequence | seqan3::views::translate_single | seqan3::ranges::to<seqan3::aa27_vector>();
         random_access_impl(state, translated_aa_sequence, access_positions);
     }
     else
@@ -113,7 +113,7 @@ void copy_impl(benchmark::State & state, std::vector<seqan3::dna4> const & dna_s
     {
         seqan3::aa27_vector translated_aa_sequence{};
         benchmark::DoNotOptimize(translated_aa_sequence =
-                                     dna_sequence | adaptor | seqan3::views::to<seqan3::aa27_vector>);
+                                     dna_sequence | adaptor | seqan3::ranges::to<seqan3::aa27_vector>());
     }
 }