+++ test/unit/search/fm_index_cursor/bi_fm_index_cursor_test.cpp
@@ -13,7 +13,7 @@
 #include <seqan3/alphabet/nucleotide/dna5.hpp>
 #include <seqan3/alphabet/views/char_to.hpp>
 #include <seqan3/search/fm_index/bi_fm_index.hpp>
-#include <seqan3/utility/views/to.hpp>
+#include <seqan3/utility/range/to.hpp>
 
 #include "bi_fm_index_cursor_test_template.hpp"
 
@@ -25,13 +25,13 @@ struct bi_fm_index_cursor_test : public ::testing::Test
     using alphabet_type = typename index_type::alphabet_type;
     using text_type = std::vector<alphabet_type>;
 
-    static constexpr auto convert = seqan3::views::char_to<alphabet_type> | seqan3::views::to<text_type>;
+    static constexpr auto convert = seqan3::views::char_to<alphabet_type> | seqan3::ranges::to<text_type>();
 
     text_type text{convert(std::string_view{"ACGGTAGGACGTAGC"})};
     text_type text1{convert(std::string_view{"AACGATCGGA"})};
 
-    text_type rev_text1 = text | std::views::reverse | seqan3::views::to<text_type>;
-    text_type rev_text2 = text1 | std::views::reverse | seqan3::views::to<text_type>;
+    text_type rev_text1 = text | std::views::reverse | seqan3::ranges::to<text_type>();
+    text_type rev_text2 = text1 | std::views::reverse | seqan3::ranges::to<text_type>();
 
     text_type pattern1{convert(std::string_view{"CAG"})};
     text_type pattern2{convert(std::string_view{"TT"})};