+++ test/unit/search/helper.hpp
@@ -14,7 +14,7 @@
 #include <seqan3/core/debug_stream/debug_stream_type.hpp>
 #include <seqan3/search/fm_index/bi_fm_index_cursor.hpp>
 #include <seqan3/search/fm_index/fm_index_cursor.hpp>
-#include <seqan3/utility/views/to.hpp>
+#include <seqan3/utility/range/to.hpp>
 
 namespace seqan3
 {
@@ -34,7 +34,7 @@ inline debug_stream_type<char_t> & operator<<(debug_stream_type<char_t> & s,
 template <typename result_range_t>
 std::vector<std::ranges::range_value_t<result_range_t>> uniquify(result_range_t && result_range)
 {
-    auto unique_res = result_range | views::to<std::vector>;
+    auto unique_res = result_range | ranges::to<std::vector>();
     std::sort(unique_res.begin(), unique_res.end());
     unique_res.erase(std::unique(unique_res.begin(), unique_res.end()), unique_res.end());
     return unique_res;